# openapi/ferrumgate-control-api.v1.yaml
openapi: 3.1.0
info:
  title: FerrumGate Control API
  version: 0.1.0
  description: REST control plane API for FerrumGate v1.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/intents/compile:
    post:
      summary: Compile an IntentEnvelope
      operationId: compileIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentCompileRequest'
      responses:
        '201':
          description: Intent compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentCompileResponse'
        '400':
          $ref: '#/components/responses/ApiError'
...
  /v1/proposals/{proposal_id}/evaluate:
    post:
      summary: Evaluate proposal via policy
      operationId: evaluateProposal
      parameters:
        - $ref: '#/components/parameters/ProposalId'
      responses:
        '200':
          description: Evaluation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateProposalResponse'
        '404':
          $ref: '#/components/responses/ApiError'
  /v1/capabilities/mint:
    post:
      summary: Mint a capability lease
      operationId: mintCapability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapabilityMintRequest'
      responses:
        '201':
          description: Capability minted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityMintResponse'
        '400':
          $ref: '#/components/responses/ApiError'
        '403':
          $ref: '#/components/responses/ApiError'
...
  /v1/capabilities/{capability_id}/revoke:
    post:
      summary: Revoke capability
      operationId: revokeCapability
      parameters:
        - $ref: '#/components/parameters/CapabilityId'
      responses:
        '200':
          description: Capability revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /v1/executions/authorize:
    post:
      summary: Authorize execution from proposal + capability
      operationId: authorizeExecution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeExecutionRequest'
      responses:
        '200':
          description: Execution authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizeExecutionResponse'
        '403':
          $ref: '#/components/responses/ApiError'
        '422':
          $ref: '#/components/responses/ApiError'
  /v1/executions/{execution_id}/prepare:
    post:
      summary: Prepare rollback / preflight checks
      operationId: prepareExecution
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
      responses:
        '200':
          description: Prepare result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrepareExecutionResponse'
        '404':
          $ref: '#/components/responses/ApiError'
...
components:
  parameters:
...
    ProposalId:
      name: proposal_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CapabilityId:
      name: capability_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ExecutionId:
      name: execution_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [code, message, correlation_id, retriable, details]
      properties:
        code:
          type: string
          enum:
            - ValidationError
            - NotFound
            - PolicyDenied
            - ApprovalRequired
            - CapabilityExpired
            - CapabilityRevoked
            - IntegrityMismatch
            - RollbackUnsupported
            - AdapterFailure
            - Conflict
            - Internal
        message:
          type: string
        correlation_id:
          type: string
        retriable:
          type: boolean
        details:
          type: object
...
    IntentCompileRequest:
      type: object
      required: [principal_id, title, goal, raw_inputs, requested_resource_scope, metadata]
      properties:
        principal_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
          nullable: true
        channel_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        goal:
          type: string
        agent_plan_summary:
          type: string
          nullable: true
        trusted_context:
          type: object
        raw_inputs:
          type: array
          items:
            $ref: '#/components/schemas/IntentInputRef'
        requested_resource_scope:
          type: array
          items:
            type: object
        requested_risk_tier:
          type: string
          nullable: true
        metadata:
          type: object
    IntentInputRef:
      type: object
      required: [source_id, source_type, trust_labels, sensitivity_labels, summary]
      properties:
        source_id:
          type: string
        source_type:
          type: string
        trust_labels:
          type: array
          items:
            type: string
        sensitivity_labels:
          type: array
          items:
            type: string
        summary:
          type: string
        event_id:
          type: string
          format: uuid
          nullable: true
    IntentCompileResponse:
      type: object
      required: [envelope, warnings]
      properties:
        envelope:
          $ref: '#/components/schemas/IntentEnvelope'
        warnings:
          type: array
          items:
            type: string
    IntentEnvelope:
      type: object
      required:
        - intent_id
        - principal_id
        - title
        - goal
        - normalized_goal
        - allowed_outcomes
        - forbidden_outcomes
        - resource_scope
        - risk_tier
        - approval_mode
        - default_rollback_class
        - time_budget
        - trust_context
        - status
        - created_at
        - expires_at
      properties:
        intent_id:
          type: string
          format: uuid
        principal_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
          nullable: true
        channel_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        goal:
          type: string
        normalized_goal:
          type: string
        allowed_outcomes:
          type: array
          items:
            type: object
        forbidden_outcomes:
          type: array
          items:
            type: object
        resource_scope:
          type: array
          items:
            type: object
        risk_tier:
          type: string
        approval_mode:
          type: string
        default_rollback_class:
          type: string
        time_budget:
          type: object
        trust_context:
          type: object
        derived_from_event_ids:
          type: array
          items:
            type: string
            format: uuid
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        status:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...
    EvaluateProposalResponse:
      type: object
      required: [decision, reason, matched_rule_ids, warnings]
      properties:
        decision:
          type: string
        reason:
          type: string
        matched_rule_ids:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
    CapabilityMintRequest:
      type: object
      required: [intent_id, proposal_id, tool_binding, resource_bindings, argument_constraints, taint_budget, requested_ttl_secs, metadata]
      properties:
        intent_id:
          type: string
          format: uuid
        proposal_id:
          type: string
          format: uuid
        tool_binding:
          type: object
        resource_bindings:
          type: array
          items:
            type: object
        argument_constraints:
          type: array
          items:
            type: object
        taint_budget:
          type: object
        approval_binding:
          type: object
          nullable: true
        requested_ttl_secs:
          type: integer
        metadata:
          type: object
    CapabilityMintResponse:
      type: object
      required: [lease, warnings]
      properties:
        lease:
          $ref: '#/components/schemas/CapabilityLease'
        warnings:
          type: array
          items:
            type: string
    CapabilityLease:
      type: object
      required:
        - capability_id
        - intent_id
        - proposal_id
        - tool_binding
        - resource_bindings
        - argument_constraints
        - taint_budget
        - issued_by
        - policy_bundle_id
        - status
        - issued_at
        - expires_at
      properties:
        capability_id:
          type: string
          format: uuid
        intent_id:
          type: string
          format: uuid
        proposal_id:
          type: string
          format: uuid
        tool_binding:
          type: object
        resource_bindings:
          type: array
          items:
            type: object
        argument_constraints:
          type: array
          items:
            type: object
        taint_budget:
          type: object
        approval_binding:
          type: object
          nullable: true
        issued_by:
          type: string
        policy_bundle_id:
          type: string
          format: uuid
        tool_manifest_id:
          type: string
          format: uuid
          nullable: true
        manifest_hash:
          type: string
          nullable: true
        status:
          type: string
        issued_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
    AuthorizeExecutionRequest:
      type: object
      required: [proposal_id, capability_id, dry_run]
      properties:
        proposal_id:
          type: string
          format: uuid
        capability_id:
          type: string
          format: uuid
        dry_run:
          type: boolean
    AuthorizeExecutionResponse:
      type: object
      required: [execution, warnings]
      properties:
        execution:
          $ref: '#/components/schemas/ExecutionRecord'
        warnings:
          type: array
          items:
            type: string
    ExecutionRecord:
      type: object
      required: [execution_id, proposal_id, intent_id, capability_id, decision, state, started_at]
      properties:
        execution_id:
          type: string
          format: uuid
        proposal_id:
          type: string
          format: uuid
        intent_id:
          type: string
          format: uuid
        capability_id:
          type: string
          format: uuid
        rollback_contract_id:
          type: string
          format: uuid
          nullable: true
        decision:
          type: string
        state:
          type: string
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        result_digest:
          type: string
          nullable: true
        metadata:
          type: object
    PrepareExecutionResponse:
      type: object
      required: [execution_id, prepared, warnings]
      properties:
        execution_id:
          type: string
          format: uuid
        prepared:
          type: boolean
        rollback_contract:
          $ref: '#/components/schemas/RollbackContract'
        warnings:
          type: array
          items:
            type: string
    RollbackContract:
      type: object
      required:
        - contract_id
        - intent_id
        - proposal_id
        - execution_id
        - action_type
        - rollback_class
        - adapter_key
        - target
        - prepare_checks
        - verify_checks
        - compensation_plan
        - auto_commit
        - state
        - created_at
      properties:
        contract_id:
          type: string
          format: uuid
        intent_id:
          type: string
          format: uuid
        proposal_id:
          type: string
          format: uuid
        execution_id:
          type: string
          format: uuid
        action_type:
          type: string
        rollback_class:
          type: string
        adapter_key:
          type: string
        target:
          type: object
        prepare_checks:
          type: array
          items:
            type: object
        verify_checks:
          type: array
          items:
            type: object
        compensation_plan:
          type: array
          items:
            type: object
        auto_commit:
          type: boolean
        state:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
...